delete all the stuff we are not using anymore
