+++ lib/api.js
@@ -1,13 +1,5 @@
-var pull = require('pull-stream')
-
 function isFunction (f) {
   return 'function' === typeof f
-}
-
-function toBuffer(v) {
-  if (Buffer.isBuffer(v)) return v
-  if (typeof v == 'string') return new Buffer(v, 'hex')
-  return new Buffer(v)
 }
 
 function each(obj, iter) {
@@ -59,4 +51,4 @@ function api (server) {
 
 }
 
-iexports = module.exports = api
+exports = module.exports = api
